+++ components/empty/src/empty.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import { computed } from 'vue'
+import { computed, useSlots } from 'vue'
 import TnIcon from '../../icon/src/icon.vue'
 import { emptyDefaultTips } from '../libs'
 import { emptyProps } from './empty'
@@ -7,7 +7,16 @@ import { useEmptyCustomStyle } from './composables'
 
 const props = defineProps(emptyProps)
 
-const { ns, emptyClass, emptyStyle, iconTextStyle } = useEmptyCustomStyle(props)
+const slots = useSlots()
+// 判断是否自定义icon或者tips的插槽
+const customIconContent = computed<boolean>(() => !!slots?.icon ?? false)
+const customTipsContent = computed<boolean>(() => !!slots?.tips ?? false)
+
+const { ns, emptyClass, emptyStyle, iconTextStyle } = useEmptyCustomStyle(
+  props,
+  customIconContent,
+  customTipsContent
+)
 
 // 根据类型获取默认提示
 const emptyTips = computed<string>(() =>
@@ -18,7 +27,10 @@ const emptyTips = computed<string>(() =>
 <template>
   <view :class="[emptyClass]" :style="emptyStyle">
     <!-- 图标 -->
-    <view :class="[ns.e('icon')]" :style="iconTextStyle('icon')">
+    <view
+      :class="[ns.e('icon'), ns.is('custom', customIconContent)]"
+      :style="iconTextStyle('icon')"
+    >
       <slot name="icon">
         <TnIcon :name="`empty-${mode}`" />
       </slot>
@@ -26,7 +38,7 @@ const emptyTips = computed<string>(() =>
     <!-- 提示文字 -->
     <view
       v-if="showTips || $slots.tips"
-      :class="[ns.e('tips')]"
+      :class="[ns.e('tips'), ns.is('custom', customTipsContent)]"
       :style="iconTextStyle('tips')"
     >
       <slot name="tips">