+++ components/swiper/src/composables/index.ts
@@ -1,0 +1,2 @@
+export * from './swiper-custom'
+export * from './use-swiper'