+++ components/swiper/src/instance.ts
@@ -1,0 +1,5 @@
+import type Swiper from './swiper.vue'
+import type SwiperItem from './swiper-item.vue'
+
+export type TnSwiperInstance = InstanceType<typeof Swiper>
+export type TnSwiperItemInstance = InstanceType<typeof SwiperItem>