+++ lib/index.ts
@@ -5,5 +5,4 @@ export * from '../hooks'
 export * from './make-installer'
 
 export const install = installer.install
-export const version = installer.version
 export default installer