+++ CHANGELOG.md
@@ -1047,3 +1047,41 @@ _2023-07-12_
 _2023-07-12_
 
 #### 修改`step`组件名称为`steps-item`
+
+
+### 1.0.6-beta.1
+
+_2023-07-13_
+
+#### 新增
+
+- `TnImageUpload`组件新增`auto-remove-faild-file`属性，当上传失败时是否自动移除文件
+
+#### Fixed
+
+- 修复`TnLazyLoad`有默认背景，在现在透明图片时带了背景颜色的问题
+- 修复`TnRate`在一定条件下会返回Nan的问题
+- 修复`TnRate`配合`TnPopup`会获取容器信息失效的问题
+- 修复`TnNumberBox`在disabled状态下能点击的问题
+- 修复`TnNumberBox`在输入模式下输入的值大于最大值时回显失败的问题
+- 修复`TnDateTimePicker`在`time`下默认值不滚动到对应位置的问题
+
+
+### 1.0.6-beta.2
+
+_2023-07-14_
+
+#### 新增
+
+- 新增`TnSwiper`组件
+
+
+### 1.0.6-beta.3
+
+_2023-07-15_
+
+#### Fixed
+
+- 修复`TnSearchBox` v-model 无法更新内部值的问题
+- 修复`TnSearchBox` 在内容为空时点击search无法触发事件的问题
+- 修复`TnEmpty`在自定义内容时无法正常显示的问题